Add pocketsphinx speech toolkit
